Add missing `MetadataFields` in `SearchResult`
